add reconstruction method benchmark tutorial
